% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcollect.R
\name{cluster}
\alias{cluster}
\title{cluster function}
\usage{
cluster(x)
}
\arguments{
\item{x}{name from original dataframe which specifies the cluster or object id.}
}
\value{
No return value, called to construct \code{formula}.
}
\description{
{Specify cluster id for clustered data or object id for recurrent data in the input \code{formula}.}
}
