% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.fpen}
\alias{coef.fpen}
\title{retrieve the coefficients under given tuning parameter}
\usage{
\method{coef}{fpen}(object, ...)
}
\arguments{
\item{object}{Object with class "fpen", generated from \code{frailtyMMpen}.}

\item{...}{Ignored}
}
\value{
A vector of estimated parameters.
}
\description{
retrieve the coefficients under given tuning parameter
}
\details{
Without given a specific tune value, the coefficients with minimum BIC is returned. If \code{tune=a},
the coefficient is computed using linear interpolation of the result from the coefficients estimated from the run of regularization path.
Thus, \code{a} should between the minimum and maximum value of the tuning parameter sequences used for the model fitting.
}
