% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcollect.R
\name{event}
\alias{event}
\title{event function}
\usage{
event(x)
}
\arguments{
\item{x}{name from original dataframe which specifies the event id.}
}
\description{
{Specify event id for multi-event data in the input \code{formula}.}
}
