% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtyMM.R
\name{frailtyMM}
\alias{frailtyMM}
\title{Fitting frailty models with clustered, multi-event and recurrent data using MM algorithm}
\usage{
frailtyMM(
  formula,
  data,
  frailty = "LogN",
  power = NULL,
  tol = 1e-05,
  maxit = 200,
  ...
)
}
\arguments{
\item{formula}{Formula where the left hand side is an object of the type \code{Surv}
and the right hand side contains the variables and additional specifications. 
\code{+cluster()} function specify the group id for clustered data or individual id for recurrent data.
\code{+event()} function specify the event id for multi-event data (only two events are allowed).}

\item{data}{The \code{data.frame} where the formula argument can be evaluated.}

\item{frailty}{The frailty used for model fitting. The default is "LogN", other choices are
"InvGauss", "Gamma" and "PVF". (Note that the computation time for PVF family will be slow 
due to the non-explicit expression of likelihood function)}

\item{power}{The power used if PVF frailty is applied.}

\item{tol}{The tolerance level for convergence.}

\item{maixt}{Maximum iterations for MM algorithm.}
}
\value{
An object of class \code{fmm} that contains the following fields:
\item{coef}{coeficient estimated from a specific model.}
\item{est.tht}{frailty parameter estimated from a specific model.}
\item{lambda}{frailty for each observation estimated from a specific model.}
\item{likelihood}{The observed log-likelihood given estimated parameters.}
\item{input}{input data.}
\item{frailty}{frailty used for model fitting.}
\item{power}{power used for model fitting is PVF frailty is applied.}
\item{iter}{total number of iterations.}
\item{convergence}{convergence threshold.}
\item{formula}{formula applied as input.}
\item{coefname}{name of each coeficient from input.}
\item{id}{id for individuals or clusters.}
\item{N}{total number of observarions.}
\item{a}{total number of individuals or clusters.}
\item{datatype}{model used for fitting.}
}
\description{
{
This formula is used to fit the non-penalized regression. 3 types of the models can be fitted, shared frailty model where
hazard rate of \eqn{j^{th}} object in \eqn{i^{th}} cluster is
\deqn{\lambda_{ij}(t|\omega_i) = \lambda_0(t) \omega_i \exp(\beta' \mathbf{X_{ij}}).}
The multi-event frailty model with different baseline hazard of different event and the hazard rate of \eqn{j^{th}} event for individual \eqn{i^{th}} is 
\deqn{\lambda_{ij}(t|\omega_i) = \lambda_{0j}(t) \omega_i \exp(\beta' \mathbf{X_{ij}}).}
}
The recurrent event model where the \eqn{j^{th}} event of individual \eqn{i} has observed feature \eqn{\mathbf{X_{ij}}},
\deqn{\lambda_{ij}(t|\omega_i) = \lambda_0(t) \omega_i \exp(\beta' \mathbf{X_{ij}}).}
}
\details{
To run the shared frailty model, \code{Surv(tstop, status)} formula should be applied along with \code{+cluster()} to specify the
corresponding clusters, if you want to run the frailty model without shared frailty, you should also use \code{+cluster()} for individual id 
to treat it as a special case of shared frailty model with only 1 object in each cluster. To run the multi-event model, 
\code{Surv(tstop, status)} formula should be applied along with \code{+event()} to specify the corresponding events. To run the recurrent event model, 
\code{Surv(tstart, tstop, status)} formula should be applied along with \code{+cluster()} where the cluster here denotes the individual id and
each individual may have many observed events at different time points.

The default frailty will be log-normal frailty, in order to fit other frailty models, simply set parameter \code{frailty} as "InvGauss", "Gamma" or "PVF",
the parameter \code{power} is only used when \code{frailty}=PVF and since the likelihood of PVF (tweedie) distribution is approximated using 
\code{Tweedie} function from package mgcv, 1<\code{power}<2.
}
