% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{hdCLdata}
\alias{hdCLdata}
\title{Simulated High-dimensional Clustered data}
\format{
An object of class \code{data.frame} with 200 rows and 503 columns.
}
\usage{
data(hdCLdata)
}
\description{
This data include 50 clusters with 4 objects, a total of 200 events are recorded.
500 covariates can be used for model fitting. The data is generated from a
gamma frailty model with coefficients (5,5,5,5,...,0) and frailty variance 1.
}
\examples{
data(hdCLdata)
}
\keyword{datasets}
