% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fpen}
\alias{print.fpen}
\title{print a penalized regression object}
\usage{
\method{print}{fpen}(x, ...)
}
\arguments{
\item{x}{Object with class "fpen" fitted by function \code{frailtyMMpen}.}

\item{...}{Ignored}
}
\value{
No return value, called to print the summary for penalized regression.
}
\description{
Print the summary of a non-penalized regression fitted by any model with function \code{frailtyMMpen}.
The first column is the tuning parameter sequence, the second column is the degree of freedom and the third column is the BIC.
}
\seealso{
\code{\link{frailtyMMpen}}
}
