% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdataCL}
\alias{simdataCL}
\title{Simulated Clustered data}
\format{
An object of class \code{data.frame} with 500 rows and 33 columns.
}
\usage{
data(simdataCL)
}
\description{
This data include 50 clusters with 10 objects, a total of 100 events are recorded.
30 temporal covariates can be used for model fitting. The data is generated from a
gamma frailty model with coefficients (1,2,3,4,0,...,0) and frailty variance 1.
}
\examples{
data(simdataCL)
}
\keyword{datasets}
