% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdataRE}
\alias{simdataRE}
\title{Simulated Recurrent Event data}
\format{
An object of class \code{data.frame} with 706 rows and 34 columns.
}
\usage{
data(simdataRE)
}
\description{
This data include 50 individuals with recurrent observation of events, a total of 706 events are recorded.
30 temporal covariates can be used for model fitting. The data is generated from a
gamma frailty model with coefficients (1,2,3,4,0,...,0) and frailty variance 1.
}
\examples{
data(simdataRE)
}
\keyword{datasets}
