% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fmm}
\alias{summary.fmm}
\title{Provide the summary for the model fitting}
\usage{
##S3 method for class "fmm"
}
\arguments{
\item{object}{Object with class "fmm", generated from \code{frailtyMM}.}

\item{...}{Ignored}
}
\description{
Provide the summary for the model fitting
}
\details{
the summary for the model of frailtyMM.
The standard error and p-value of estimated parameters are based on Fisher Information martix.
}
