% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.fpen}
\alias{coef.fpen}
\title{retreive the coefficence under given tuning parameter}
\usage{
\method{coef}{fpen}(x, tune = NULL)
}
\description{
retreive the coefficence under given tuning parameter
}
\details{
Without given a specific tune value, the coefficenct with minimum BIC is returned. If \code{tune=a},
the coefficient is computed using linear interpolation of the result from the coefficents estimated from the run of regularization path.
Thus, \code{a} should between the minimum and maximum value of the tuning parameter sequences used for the model fitting.
}
