% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcollect.R
\name{event}
\alias{event}
\title{#' cluster function
#'
#' @param x name from original dataframe which specifies the cluster or object id.
#' @description {Specify cluster id for clustered data or object id for recurrent data in the input \code{formula}.}
#' @return No return value, called to construct \code{formula}.
#' @export
cluster <- function(x) {
  x
}
event function}
\usage{
event(x)
}
\arguments{
\item{x}{name from original dataframe which specifies the event id.}
}
\value{
No return value, called to construct \code{formula}.
}
\description{
{Specify event id for multi-event data in the input \code{formula}.}
}
