% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtyMM.R
\name{frailtyMM}
\alias{frailtyMM}
\title{Fitting frailty models with clustered, multi-event and recurrent data using MM algorithm}
\usage{
frailtyMM(
  formula,
  data,
  frailty = "LogN",
  power = NULL,
  tol = 1e-06,
  maxit = 200,
  ...
)
}
\arguments{
\item{formula}{Formula where the left hand side is an object of the type \code{Surv}
and the right hand side contains the variables and additional specifications. 
\code{+cluster()} function specify the group id for clustered data or individual id for recurrent data.
\code{+event()} function specify the event id for multi-event data (only two events are allowed).}

\item{data}{The \code{data.frame} where the formula argument can be evaluated.}

\item{frailty}{The frailty used for model fitting. The default is "LogN", other choices are
"InvGauss", "Gamma" and "PVF". (Note that the computation time for PVF family will be slow 
due to the non-explicit expression of likelihood function)}

\item{power}{The power used if PVF frailty is applied.}

\item{tol}{The tolerance level for convergence.}

\item{maixt}{Maximum iterations for MM algorithm.}
}
\description{
Fitting frailty models with clustered, multi-event and recurrent data using MM algorithm
}
