% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frailtyMMpen.R
\name{frailtyMMpen}
\alias{frailtyMMpen}
\title{Fitting penalized frailty models with clustered, multi-event and recurrent data using MM algorithm}
\usage{
frailtyMMpen(
  formula,
  data,
  frailty = "LogN",
  power = NULL,
  penalty = "LASSO",
  tune = NULL,
  tol = 1e-05,
  maxit = 200
)
}
\arguments{
\item{formula}{Formula where the left hand side is an object of the type \code{Surv}
and the right hand side contains the variables and additional specifications. 
\code{+cluster()} function specify the group id for clustered data or individual id for recurrent data.
\code{+event()} function specify the event id for multi-event data (only two events are allowed).}

\item{data}{The \code{data.frame} where the formula argument can be evaluated.}

\item{frailty}{The frailty used for model fitting. The default is "LogN", other choices are
"InvGauss", "Gamma" and "PVF". (Note that the computation time for PVF family will be slow 
due to the non-explicit expression of likelihood function)}

\item{power}{The power used if PVF frailty is applied.}

\item{penalty}{The penalty used for regulization, the default is "LASSO", other choices are "MCP" and "SCAD".}

\item{tune}{The sequence of tuning parameters provided by user. If not provided, the default grid will be applied.}

\item{tol}{The tolerance level for convergence.}

\item{maixt}{Maximum iterations for MM algorithm.}
}
\value{
An object of class \code{fmm} that contains the following fields:
\item{coef}{matrix of coeficient estimated from a specific model.}
\item{est.tht}{vector of frailty parameter estimated from a specific model.}
\item{lambda}{list of frailty for each observation estimated from a specific model.}
\item{likelihood}{vector of the observed log-likelihood given estimated parameters.}
\item{BIC}{vector of the BIC given estimated parameters.}
\item{tune}{vector of tuning parameters used for penalized regression.}
\item{tune.min}{tuning parameter where minimal of BIC is obtained.}
\item{convergence}{convergence threshold.}
\item{y}{input stopping time.}
\item{X}{input covariate matrix.}
\item{d}{input censoring indicator.}
}
\description{
{This formula is used to fit the non-penalized regression. 3 types of the models can be fitted similar to the function
\code{frailtyMM}. In addition, variable selection can be done by three types of penalty, LASSO, MCP and SCAD.
}
}
\details{
Without a given \code{tune}, the default sequence of tuning parameters are used to provide the regularization path, 
the \eqn{a} used for MCP is 3 and SCAD is 3.7. The formula is same as the input for function \code{frailtyMM}.
}
\seealso{
\code{\link{frailtyMM}}
}
