% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{hdCLdata}
\alias{hdCLdata}
\title{Simulated High-dimensional Clustered data}
\format{
An object of class \code{data.frame} with 80 rows and 203 columns.
}
\usage{
data(hdCLdata)
}
\description{
This data include 40 clusters with 2 objects, a total of 80 events are recorded.
200 covariates can be used for model fitting. The data is generated from a
gamma frailty model with coefficients (4,4,4,4,4,...,0) and frailty variance 1.
}
\examples{
data(hdCLdata)
}
\keyword{datasets}
