% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fmm}
\alias{plot.fmm}
\title{Plot the regularization path}
\usage{
##S3 method for class "fpen"
}
\arguments{
\item{x}{Object with class "fpen"}

\item{...}{Further arguments pass to or from other methods}
}
\description{
Plot the whole regularization path run by frailtyMMpen
}
