% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fmm}
\alias{print.fmm}
\title{print a non-penalized regression object}
\usage{
\method{print}{fmm}(x, ...)
}
\arguments{
\item{x}{Object with class "fmm" fitted by function \code{frailtyMM}.}

\item{...}{Ignored}
}
\description{
Print the summary of a non-penalized regression fitted by any model with function \code{frailtyMM}
}
\seealso{
\code{\link{frailtyMM}}
}
